requirements.txt
# Core dependencies
httpx>=0.24.0
python-dotenv>=1.0.0
mcp>=0.1.0
fastmcp>=0.1.0
python-dateutil>=2.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Optional: Redis for distributed caching
redis>=4.6.0